Fixed the stack, and printing

+++ src/stack.rs
@@ -3,7 +3,7 @@ use std::fmt::Debug;
 //Stack is a data structure that stores a list of elements(of given type) in a Last In First Out order
 #[derive(Debug)]
 pub struct Stack{
-    stack: Vec<i32>,
+    stack: Vec<i64>,
 }
 
 //Methods
@@ -17,12 +17,12 @@ impl<> Stack<> {
     }
 
     //Push an element to the top of the stack
-    pub fn push(&mut self, item: i32) {
+    pub fn push(&mut self, item: i64) {
         self.stack.push(item);
     }
 
     //Pop the top element from the stack
-    pub fn pop(&mut self) -> Option<i32> {
+    pub fn pop(&mut self) -> Option<i64> {
         self.stack.pop()
     }
 
@@ -38,7 +38,7 @@ impl<> Stack<> {
     }
 
     //Return the next element in the stack without removing it
-    pub fn peek(&self) -> Option<&i32> {
+    pub fn peek(&self) -> Option<&i64> {
         self.stack.last()
     }
